[UPD] - Update variable types and access modifiers across multiple files for improved code clarity and compliance with Explicit API mode, including changes in DebugLogService, PlayerScreen, SettingsViewModel, TopicViewModel, and various Torrent-related classes.

+++ android/app/src/main/kotlin/com/jabook/app/jabook/compose/data/debug/DebugLogService.kt
@@ -328,7 +328,7 @@ public class DebugLogService
                         } ?: emptyArray()
 
                     val now = System.currentTimeMillis()
-                    val maxAge: Int = 7 * 24 * 60 * 60 * 1000L // 7 days
+                    val maxAge: Long = 7 * 24 * 60 * 60 * 1000L // 7 days
                     var deletedCount: Int = 0
                     logFiles.forEach { file ->
                         if (now - file.lastModified() > maxAge) {
